migrate remaining tools and classes out of SolrCLI
