+++ solr/core/src/java/org/apache/solr/util/cli/HealthcheckTool.java
@@ -49,7 +49,7 @@ import org.noggit.JSONWriter;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-public class HealthcheckTool extends SolrCLI.SolrCloudTool {
+public class HealthcheckTool extends SolrCloudTool {
   private static final Logger log = LoggerFactory.getLogger(MethodHandles.lookup().lookupClass());
 
   enum ShardState {
