+++ solr/core/src/java/org/apache/solr/util/cli/StatusTool.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
- 
+
 package org.apache.solr.util.cli;
 
 import java.io.PrintStream;
