+++ solr/core/src/test/org/apache/solr/util/TestSolrCLIRunExample.java
@@ -48,6 +48,8 @@ import org.apache.solr.cloud.MiniSolrCloudCluster;
 import org.apache.solr.common.SolrInputDocument;
 import org.apache.solr.embedded.JettyConfig;
 import org.apache.solr.embedded.JettySolrRunner;
+import org.apache.solr.util.cli.DeleteTool;
+import org.apache.solr.util.cli.RunExampleTool;
 import org.junit.After;
 import org.junit.AfterClass;
 import org.junit.BeforeClass;
@@ -365,7 +367,7 @@ public class TestSolrCLIRunExample extends SolrTestCaseJ4 {
       RunExampleExecutor executor = new RunExampleExecutor(stdoutSim);
       closeables.add(executor);
 
-      SolrCLI.RunExampleTool tool = new SolrCLI.RunExampleTool(executor, System.in, stdoutSim);
+      RunExampleTool tool = new RunExampleTool(executor, System.in, stdoutSim);
       try {
         int status =
             tool.runTool(
@@ -488,7 +490,7 @@ public class TestSolrCLIRunExample extends SolrTestCaseJ4 {
     RunExampleExecutor executor = new RunExampleExecutor(stdoutSim);
     closeables.add(executor);
 
-    SolrCLI.RunExampleTool tool = new SolrCLI.RunExampleTool(executor, userInputSim, stdoutSim);
+    RunExampleTool tool = new RunExampleTool(executor, userInputSim, stdoutSim);
     try {
       tool.runTool(
           SolrCLI.processCommandLineArgs(
@@ -554,7 +556,7 @@ public class TestSolrCLIRunExample extends SolrTestCaseJ4 {
     }
 
     // delete the collection
-    SolrCLI.DeleteTool deleteTool = new SolrCLI.DeleteTool(stdoutSim);
+    DeleteTool deleteTool = new DeleteTool(stdoutSim);
     String[] deleteArgs = new String[] {"-name", collectionName, "-solrUrl", solrUrl};
     deleteTool.runTool(
         SolrCLI.processCommandLineArgs(
@@ -603,7 +605,7 @@ public class TestSolrCLIRunExample extends SolrTestCaseJ4 {
 
     DefaultExecutor executor = new DefaultExecutor();
 
-    SolrCLI.RunExampleTool tool = new SolrCLI.RunExampleTool(executor, System.in, stdoutSim);
+    RunExampleTool tool = new RunExampleTool(executor, System.in, stdoutSim);
     int code =
         tool.runTool(
             SolrCLI.processCommandLineArgs(
