+++ solr/core/src/test/org/apache/solr/util/cli/AuthToolTest.java
@@ -25,13 +25,11 @@ import java.nio.file.Path;
 import org.apache.commons.cli.CommandLine;
 import org.apache.commons.io.file.PathUtils;
 import org.apache.solr.cloud.SolrCloudTestCase;
-import org.apache.solr.util.SolrCLI;
 import org.junit.After;
 import org.junit.Before;
 import org.junit.BeforeClass;
 import org.junit.Test;
 
-/** Unit test for SolrCLI's AuthTool */
 public class AuthToolTest extends SolrCloudTestCase {
   private Path dir;
 
@@ -82,7 +80,7 @@ public class AuthToolTest extends SolrCloudTestCase {
 
   private int runTool(String[] args) throws Exception {
     Tool tool = findTool(args);
-    assertTrue(tool instanceof SolrCLI.AuthTool);
+    assertTrue(tool instanceof AuthTool);
     CommandLine cli = parseCmdLine(tool.getName(), args, tool.getOptions());
     return tool.runTool(cli);
   }
